+++ assets/appFramework/vue/generate/component/src/{GEN_NAME_SNAKE}.ts
@@ -1,12 +1,13 @@
 /**
  * {GEN_NAME_HUMAN} component class.
  */
-import Vue from "vue";
+import {SYNTHETIC_IMPORT}Vue from "vue";
 import Component from "vue-class-component";
 import "./{GEN_NAME_SNAKE}.css";
+import {SYNTHETIC_IMPORT}importedTemplate from "./{GEN_NAME_SNAKE}.vue";
 
 @Component({
-    "template": "url:./{GEN_NAME_SNAKE}.vue"
+    template: importedTemplate
 })
 export class {GEN_NAME_PASCAL} extends Vue {
 }