+++ assets/appFramework/vue/src/typescript/customTypes/vue-module.d.ts
@@ -1,0 +1,4 @@
+declare module "*.vue" {
+    const template: any;
+    export = template;
+}