+++ assets/appFramework/vue/test/e2e/src/e2eTestRunner/protractor/sourceLanguage/typescript/mochachai/app.spec.ts
@@ -19,7 +19,7 @@ describe("App", () => {
     });
 
     it("the root text is set", (done) => {
-        browser.loadAndWaitForPreactPage("/")
+        browser.loadAndWaitForVuePage("/")
             .then(() => {
                 $("#root > span").getText()
                     .then((rootContent) => {
@@ -30,7 +30,7 @@ describe("App", () => {
     });
 
     it("the font size is set", (done) => {
-        browser.loadAndWaitForPreactPage("/")
+        browser.loadAndWaitForVuePage("/")
             .then(() => {
                 $(".child").getCssValue("font-size")
                     .then((fontSize) => {