+++ assets/appFramework/vue/test/e2e/src/e2eTestRunner/webdriverio/sourceLanguage/typescript/mochachai/app.spec.ts
@@ -17,7 +17,7 @@ describe("App", () => {
 
     it("the root text is set", () => {
         return browser
-            .loadAndWaitForPreactPage("/")
+            .loadAndWaitForVuePage("/")
             .element("#root")
             .getText()
             .then((rootContent) => {
@@ -27,7 +27,7 @@ describe("App", () => {
 
     it("the font size is set", () => {
         return browser
-            .loadAndWaitForPreactPage("/")
+            .loadAndWaitForVuePage("/")
             .element(".child")
             .getCssProperty("font-size")
             .then((fontSize) => {