+++ src/pipelineSteps/applicationFramework/vue.ts
@@ -1,10 +1,12 @@
 /**
  * Pipeline step to generate scaffolding for Vue application.
  */
+import { JestConfiguration } from "src/configuration/models/jest/jestConfiguration";
 import { ArrayHelper } from "unitejs-framework/dist/helpers/arrayHelper";
 import { ObjectHelper } from "unitejs-framework/dist/helpers/objectHelper";
 import { IFileSystem } from "unitejs-framework/dist/interfaces/IFileSystem";
 import { ILogger } from "unitejs-framework/dist/interfaces/ILogger";
+import { BabelConfiguration } from "../../configuration/models/babel/babelConfiguration";
 import { EsLintConfiguration } from "../../configuration/models/eslint/esLintConfiguration";
 import { ProtractorConfiguration } from "../../configuration/models/protractor/protractorConfiguration";
 import { TsLintConfiguration } from "../../configuration/models/tslint/tsLintConfiguration";
@@ -28,17 +30,13 @@ export class Vue extends SharedAppFramework {
     }
 
     public async configure(logger: ILogger, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables, mainCondition: boolean): Promise<number> {
-        const usingGulp = super.condition(uniteConfiguration.taskManager, "Gulp");
-        if (mainCondition && usingGulp) {
-            engineVariables.buildTranspileInclude.push("const inlineVue = require(\"gulp-inline-vue\");");
-
-            engineVariables.buildTranspilePreBuild.push(".pipe(inlineVue())");
-        }
-
-        engineVariables.toggleDevDependency(["gulp-inline-vue"], mainCondition && usingGulp);
-
         engineVariables.toggleDevDependency(["unitejs-vue-protractor-plugin"], mainCondition && super.condition(uniteConfiguration.e2eTestRunner, "Protractor"));
         engineVariables.toggleDevDependency(["unitejs-vue-webdriver-plugin"], mainCondition && super.condition(uniteConfiguration.e2eTestRunner, "WebdriverIO"));
+
+        engineVariables.toggleDevDependency(["babel-plugin-transform-decorators-legacy", "babel-plugin-transform-class-properties"],
+                                            mainCondition && super.condition(uniteConfiguration.sourceLanguage, "JavaScript"));
+
+        engineVariables.toggleDevDependency(["babel-eslint"], mainCondition && super.condition(uniteConfiguration.linter, "ESLint"));
 
         engineVariables.toggleClientPackage(
             "vue",
@@ -101,6 +99,21 @@ export class Vue extends SharedAppFramework {
             mainCondition && super.condition(uniteConfiguration.bundler, "RequireJS"));
 
         engineVariables.toggleClientPackage(
+            "requirejs-text",
+            "text.js",
+            undefined,
+            undefined,
+            false,
+            "both",
+            "none",
+            false,
+            undefined,
+            { text: "requirejs-text" },
+            undefined,
+            undefined,
+            mainCondition && super.condition(uniteConfiguration.bundler, "RequireJS"));
+
+        engineVariables.toggleClientPackage(
             "systemjs-plugin-css",
             "css.js",
             undefined,
@@ -110,23 +123,54 @@ export class Vue extends SharedAppFramework {
             "none",
             false,
             undefined,
-            undefined,
-            { "*.css" : "systemjs-plugin-css" },
+            { css: "systemjs-plugin-css" },
+            { "*.css" : "css" },
             undefined,
             mainCondition &&
             (super.condition(uniteConfiguration.bundler, "Browserify") ||
              super.condition(uniteConfiguration.bundler, "SystemJSBuilder") ||
              super.condition(uniteConfiguration.bundler, "Webpack")));
 
-        if (mainCondition && super.condition(uniteConfiguration.taskManager, "Gulp") && super.condition(uniteConfiguration.bundler, "RequireJS")) {
+        engineVariables.toggleClientPackage(
+            "systemjs-plugin-text",
+            "text.js",
+            undefined,
+            undefined,
+            false,
+            "both",
+            "none",
+            false,
+            undefined,
+            { text: "systemjs-plugin-text" },
+            { "*.vue" : "text" },
+            undefined,
+            mainCondition &&
+                (super.condition(uniteConfiguration.bundler, "Browserify") ||
+                super.condition(uniteConfiguration.bundler, "SystemJSBuilder") ||
+                super.condition(uniteConfiguration.bundler, "Webpack")));
+
+        const usingGulp = super.condition(uniteConfiguration.taskManager, "Gulp");
+        if (mainCondition && usingGulp) {
             engineVariables.buildTranspileInclude.push("const replace = require(\"gulp-replace\");");
-            engineVariables.buildTranspileInclude.push("const clientPackages = require(\"./util/client-packages\");");
-            engineVariables.buildTranspilePreBuild.push(".pipe(replace(/import \"\.\\/(.*?).css\";/g,");
-            engineVariables.buildTranspilePreBuild.push("    `import \"\${clientPackages.getTypeMap(uniteConfig, \"css\", buildConfiguration.minify)}!./$1\";`))");
+
+            if (super.condition(uniteConfiguration.bundler, "RequireJS")) {
+                engineVariables.buildTranspileInclude.push("const clientPackages = require(\"./util/client-packages\");");
+
+                engineVariables.buildTranspilePreBuild.push(".pipe(replace(/import \\\"\\.\\\/(.*?).css\";/g,");
+                engineVariables.buildTranspilePreBuild.push("            `import \"\${clientPackages.getTypeMap(uniteConfig, \"css\", buildConfiguration.minify)}!./$1.css\";`))");
+                engineVariables.buildTranspilePreBuild.push("        .pipe(replace(/import(.*)importedTemplate from \\\"\\.\\\/(.*?).vue\";/g,");
+                engineVariables.buildTranspilePreBuild.push("             `import$1importedTemplate from \"text!./$2.vue\";`))");
+            }
+
+            const componentTemplateReplace = "/([\\\"\\\']?template[\\\"\\\']?:(?:.*?))(importedTemplate)/";
+            const componentTemplateWith = "`$1(/<template>([\\\\s\\\\S]*)<\\\\/template>/i.exec(importedTemplate) || [\"\", \"\"])[1]`";
+
+            engineVariables.buildTranspilePreBuild.push(`        .pipe(replace(${componentTemplateReplace}, ${componentTemplateWith}))`);
         }
 
         const esLintConfiguration = engineVariables.getConfiguration<EsLintConfiguration>("ESLint");
         if (esLintConfiguration) {
+            ObjectHelper.addRemove(esLintConfiguration, "parser", "babel-eslint", mainCondition);
             ObjectHelper.addRemove(esLintConfiguration.rules, "no-unused-vars", 1, mainCondition);
         }
 
@@ -148,9 +192,16 @@ export class Vue extends SharedAppFramework {
             ArrayHelper.addRemove(webdriverIoPlugins, "unitejs-vue-webdriver-plugin", mainCondition);
         }
 
+        const babelConfiguration = engineVariables.getConfiguration<BabelConfiguration>("Babel");
+        if (babelConfiguration) {
+            ArrayHelper.addRemove(babelConfiguration.plugins, "transform-decorators-legacy", mainCondition);
+            ArrayHelper.addRemove(babelConfiguration.plugins, "transform-class-properties", mainCondition);
+        }
+
         const typeScriptConfiguration = engineVariables.getConfiguration<TypeScriptConfiguration>("TypeScript");
         if (typeScriptConfiguration) {
             ObjectHelper.addRemove(typeScriptConfiguration.compilerOptions, "experimentalDecorators", true, mainCondition);
+            ObjectHelper.addRemove(typeScriptConfiguration.compilerOptions, "allowSyntheticDefaultImports", true, mainCondition);
         }
 
         const javaScriptConfiguration = engineVariables.getConfiguration<JavaScriptConfiguration>("JavaScript");
@@ -158,39 +209,51 @@ export class Vue extends SharedAppFramework {
             ObjectHelper.addRemove(javaScriptConfiguration.compilerOptions, "experimentalDecorators", true, mainCondition);
         }
 
+        const jestConfiguration = engineVariables.getConfiguration<JestConfiguration>("Jest");
+        if (jestConfiguration) {
+            ObjectHelper.addRemove(jestConfiguration.moduleNameMapper, "\\.vue$", "<rootDir>/test/unit/dummy.mock.js", mainCondition);
+        }
+
         return 0;
     }
 
     public async finalise(logger: ILogger, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables, mainCondition: boolean): Promise<number> {
         if (mainCondition) {
-            const sourceExtension = super.condition(uniteConfiguration.sourceLanguage, "TypeScript") ? ".ts" : ".js";
+            const isTypeScript = super.condition(uniteConfiguration.sourceLanguage, "TypeScript");
+            const sourceExtension = isTypeScript ? ".ts" : ".js";
 
-            let ret = await this.generateAppSource(logger, fileSystem, uniteConfiguration, engineVariables, [
+            const srcFiles = [
                 `app${sourceExtension}`,
                 `router${sourceExtension}`,
                 `child/child${sourceExtension}`,
                 `bootstrapper${sourceExtension}`,
                 `entryPoint${sourceExtension}`
-            ]);
+            ];
+
+            if (isTypeScript) {
+                srcFiles.push("customTypes/vue-module.d.ts");
+            }
+
+            let ret = await this.generateAppSource(logger, fileSystem, uniteConfiguration, engineVariables, srcFiles);
 
             if (ret === 0) {
                 ret = await super.generateAppHtml(logger, fileSystem, uniteConfiguration, engineVariables, ["app.vue", "child/child.vue"]);
+            }
 
-                if (ret === 0) {
-                    ret = await super.generateAppCss(logger, fileSystem, uniteConfiguration, engineVariables, [`child/child`]);
+            if (ret === 0) {
+                ret = await super.generateAppCss(logger, fileSystem, uniteConfiguration, engineVariables, [`child/child`]);
+            }
 
-                    if (ret === 0) {
-                        ret = await super.generateE2eTest(logger, fileSystem, uniteConfiguration, engineVariables, [`app.spec${sourceExtension}`]);
+            if (ret === 0) {
+                ret = await super.generateE2eTest(logger, fileSystem, uniteConfiguration, engineVariables, [`app.spec${sourceExtension}`]);
+            }
 
-                        if (ret === 0) {
-                            ret = await this.generateUnitTest(logger, fileSystem, uniteConfiguration, engineVariables, [`app.spec${sourceExtension}`, `bootstrapper.spec${sourceExtension}`], true);
+            if (ret === 0) {
+                ret = await this.generateUnitTest(logger, fileSystem, uniteConfiguration, engineVariables, [`app.spec${sourceExtension}`, `bootstrapper.spec${sourceExtension}`], true);
+            }
 
-                            if (ret === 0) {
-                                ret = await super.generateCss(logger, fileSystem, uniteConfiguration, engineVariables);
-                            }
-                        }
-                    }
-                }
+            if (ret === 0) {
+                ret = await super.generateCss(logger, fileSystem, uniteConfiguration, engineVariables);
             }
 
             return ret;